CRUD - PetOwner, Pet, Walker, Appointment

+++ Dogo.Application/Queries/Walker/DeleteWalkerQueryHandler.cs
@@ -1,20 +1,25 @@
-﻿using Dogo.Core.Repositories;
+﻿using Dogo.Core.Helpers;
+using Dogo.Core.Repositories;
 using MediatR;
 
 namespace Dogo.Application.Queries.Walker
 {
-    public class DeleteWalkerQueryHandler : IRequestHandler<DeleteWalkerQuery>
+    public class DeleteWalkerQueryHandler : IRequestHandler<DeleteWalkerQuery, Result>
     {
         private readonly IWalkerRepository _walkerRepository;
 
         public DeleteWalkerQueryHandler(IWalkerRepository walkerRepository)
             => _walkerRepository = walkerRepository;
 
-        public async Task<Unit> Handle(DeleteWalkerQuery request, CancellationToken cancellationToken)
+        public async Task<Result> Handle(DeleteWalkerQuery request, CancellationToken cancellationToken)
         {
             var walker = await _walkerRepository.GetByIdAsync(request.Id);
+            if (walker == null)
+            {
+                return Result.Failure(HttpStatusCode.NotFound, "Walker not found");
+            }
             await _walkerRepository.DeleteAsync(walker);
-            return Unit.Value;
+            return Result.Success();
         }
     }
 }
